Move event validation from events module
