+++ .vscode/settings.json
@@ -2,8 +2,6 @@
     "rust-analyzer.linkedProjects": [
         "Cargo.toml",
         "simulator/Cargo.toml",
-        "assets/blueprints/Cargo.toml",
-        "radix-engine-tests/tests/blueprints/Cargo.toml",
     ],
     "rust-analyzer.cargo.features": [
         "serde",
