+++ radix-engine/src/system/kernel_modules/events/error.rs
@@ -1,7 +1,6 @@
 use radix_engine_interface::data::scrypto::model::PackageAddress;
 use radix_engine_interface::*;
 use sbor::rust::string::String;
-use sbor::LocalTypeIndex;
 
 #[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor)]
 pub enum EventError {
@@ -10,14 +9,6 @@ pub enum EventError {
         blueprint_name: String,
         event_name: String,
     },
-    FailedToResolveLocalSchema {
-        local_type_index: LocalTypeIndex,
-    },
-    EventNameMismatch {
-        expected: String,
-        actual: String,
-    },
-    InvalidEventSchema,
     EventSchemaNotMatch(String),
     NoAssociatedPackage,
     InvalidActor,
